+++ src/main/java/com/example/demo/services/task/components/Command.java
@@ -1,4 +1,4 @@
-package com.example.demo.services.task.data;
+package com.example.demo.services.task.components;
 
 import com.example.demo.services.task.TaskException;
 
@@ -6,7 +6,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import static com.example.demo.services.task.TaskException.Type.WRONG_FORMAT;
-import static com.example.demo.services.task.data.TaskStatus.*;
+import static com.example.demo.services.task.components.TaskStatus.*;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
 public class Command {
