+++ src/main/java/com/example/demo/services/task/components/Task.java
@@ -1,11 +1,11 @@
-package com.example.demo.services.task.data;
+package com.example.demo.services.task.components;
 
 import com.example.demo.services.task.TaskException;
 
 import java.util.List;
 
 import static com.example.demo.services.task.TaskException.Type.WRONG_FORMAT;
-import static com.example.demo.services.task.data.TaskStatus.CREATED;
+import static com.example.demo.services.task.components.TaskStatus.CREATED;
 import static org.apache.commons.lang3.StringUtils.isBlank;
 
 public class Task {
